% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckHW1.R
\name{CheckHW1}
\alias{CheckHW1}
\title{Function to check results of HW1 before submitting.}
\usage{
CheckHW1(Husbands, Wives, FamilyWide, FamilyLong, DescriptivesFW,
  DescriptivesFL, DescriptivesFLTime, FamilyLong2, FamilyLong3)
}
\arguments{
\item{Husbands}{A data frame from Question #1}

\item{Wives}{A data frame from Question #1}

\item{FamilyWide}{A data frame from Question #2}

\item{FamilyLong}{A data frame from Question #3}

\item{DescriptivesFW}{A data frame from Question #4}

\item{DescriptivesFL}{A data frame from Question #3}

\item{DescriptivesFLTime}{A data frame from Question #5}

\item{FamilyLong2}{A data frame from Question #6}

\item{FamilyLong3}{A data frame from Question #7}
}
\value{
The score for HW1 if this file was submitted for credit.
}
\description{
Given the nine data frames that HW1 requires, checks each for accuracy and provides a score.
}
