% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataMath}
\alias{dataMath}
\title{Data file for examples used in EPSY 905}
\format{A data frame with 350 observations and 9 variables
\describe{
  \item{id}{id number}
  \item{hsl}{Mathematics courses taken in high school}
  \item{cc}{Credit hours in mathematics courses in college}
  \item{use}{Perceived usefulness of mathematics score}
  \item{msc}{Mathematics self concept score}
  \item{mas}{Mathematics anxiety score}
  \item{mse}{Mathematics self efficacy score}
  \item{perf}{Performance on mathematics test (number correct out of 20 items)}
  \item{female}{Dummy coded sex where female = 1}
}}
\usage{
dataMath
}
\description{
A dataset containing mathematics efficacy study data (simulated)
}
\keyword{datasets}
