% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataPerfIQ}
\alias{dataPerfIQ}
\title{Data file for examples used in EPSY 905}
\format{A data frame with 5 observations and 3 variables
\describe{
  \item{id}{id number}
  \item{iq}{IQ score}
  \item{Perf}{Job performance score}
}}
\usage{
dataPerfIQ
}
\description{
A dataset containing five observations used to show likelihoods
}
\keyword{datasets}
