% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataSAT}
\alias{dataSAT}
\title{Data file for examples used in EPSY 905}
\format{A data frame with 5 observations and 3 variables
\describe{
  \item{id}{id number}
  \item{SATV}{SAT Verbal score}
  \item{SATM}{SAT Math score}
}}
\usage{
dataSAT
}
\description{
A dataset containing SAT Math and Verbal Scores (simulated)
}
\keyword{datasets}
