% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataSexHeightWeight}
\alias{dataSexHeightWeight}
\title{Data file for examples used in EPSY 905}
\format{A data frame with 20 observations and 4 variables
\describe{
  \item{id}{id number}
  \item{sex}{sex of person}
  \item{heightIN}{height in inches}
  \item{weightLB}{weight in pounds}
}}
\usage{
dataSexHeightWeight
}
\description{
A dataset containing the heights, weights, and sex of 20 hypothetical people
}
\keyword{datasets}
