% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bvn_surface.R
\name{plot_bvn_surface}
\alias{plot_bvn_surface}
\title{A function that plots the surface of a bivariate normal distribution}
\usage{
plot_bvn_surface(meanvec, covmat, type, xlab, ylab, zlab, main)
}
\arguments{
\item{meanvec}{A matrix with 2 rows andone column that stores the mean of the bivariate normal distribution}

\item{covmat}{A 2 by 2 symmetric covariance matrix}

\item{type}{A character stating the type of plot from the plot.surface function of the fields package ("c" is a contour, "p" is a perspective/drape plot)}

\item{xlab}{A character for the x-axis label}

\item{ylab}{A character for the y-axis label}

\item{zlab}{A character for the z-axis label}

\item{main}{A character for the main label}
}
\value{
A plot of the surface of a bivariate normal with meanvec as the mean vector and covmat as the covariance matrix
}
\description{
A function that plots the surface of a bivariate normal distribution
}
\examples{
plot_bvn_surface(meanvec = matrix(c(0,0), nrow = 2, ncol = 1),
                 covmat = diag(2), type = "c", xlab = "x",
                 ylab = "y", zlab = "f(x,y)", main = "contour plot")
plot_bvn_surface(meanvec = matrix(c(0,0), nrow = 2, ncol = 1),
                 covmat = diag(2), type = "p", xlab = "x",
                 ylab = "y", zlab = "f(x,y)", main = "drape plot")
}
